Move datatype serializer/api to data_import app
Also updates migrations to match current master branch.
