+++ open_humans/urls.py
@@ -5,6 +5,7 @@ from django.contrib import admin
 from django.urls import reverse_lazy
 from django.views.generic import RedirectView, TemplateView
 
+import data_import.api_urls
 import data_import.urls
 import private_sharing.api_urls
 import private_sharing.urls
@@ -24,6 +25,7 @@ urlpatterns = [
     # Include the various APIs here
     path("api/", include(api_urls)),
     path("api/direct-sharing/", include(private_sharing.api_urls)),
+    path("api/data-management/", include(data_import.api_urls)),
     # from data_import, but alternate name as it is not specific to import
     path("data-management/", include(data_import.urls, namespace="data-management")),
     # The URLs used for the OAuth2 dance (e.g. requesting an access token)
