+++ private_sharing/serializers.py
@@ -156,25 +156,3 @@ class ProjectMemberDataSerializer(serializers.ModelSerializer):
             rep.pop("username")
 
         return rep
-
-
-class DataTypeSerializer(serializers.ModelSerializer):
-    """
-    Serialize DataTypes
-    """
-
-    class Meta:  # noqa: D101
-        model = DataType
-
-        fields = ["id", "name", "parent", "description", "source_projects"]
-
-    source_projects = serializers.SerializerMethodField()
-
-    def get_source_projects(self, obj):
-        """
-        Get projects associated with a datatype
-        """
-        projects = DataRequestProject.objects.filter(
-            registered_datatypes=obj
-        ).distinct()
-        return [project.id_label for project in projects]
